// src/types/responceTypes.ts
export interface Hit {
  recipe: Recipe
  _links: Links2
}

export interface Recipe {
  uri: string
  label: string
  image: string
  images: Images
  source: string
  url: string
  shareAs: string
  yield: number
  dietLabels: string[]
  healthLabels: string[]
  cautions: string[]
  ingredientLines: string[]
  ingredients: Ingredient[]
  calories: number
  totalWeight: number
  totalTime: number
  cuisineType: string[]
  mealType: string[]
  dishType: string[]
  totalNutrients: TotalNutrients
  totalDaily: TotalDaily
  digest: Digest[]
}

export interface Images {
  THUMBNAIL: Thumbnail
  SMALL: Small
  REGULAR: Regular
  LARGE?: Large
}

export interface Thumbnail {
  url: string
  width: number
  height: number
}

export interface Small {
  url: string
  width: number
  height: number
}

export interface Regular {
  url: string
  width: number
  height: number
}

export interface Large {
  url: string
  width: number
  height: number
}

export interface Ingredient {
  text: string
  quantity: number
  measure?: string
  food: string
  weight: number
  foodCategory: string
  foodId: string
  image?: string
}

export interface TotalNutrients {
  ENERC_KCAL: EnercKcal
  FAT: Fat
  FASAT: Fasat
  FATRN: Fatrn
  FAMS: Fams
  FAPU: Fapu
  CHOCDF: Chocdf
  FIBTG: Fibtg
  SUGAR: Sugar
  PROCNT: Procnt
  CHOLE: Chole
  NA: Na
  CA: Ca
  MG: Mg
  K: K
  FE: Fe
  ZN: Zn
  P: P
  VITA_RAE: VitaRae
  VITC: Vitc
  THIA: Thia
  RIBF: Ribf
  NIA: Nia
  VITB6A: Vitb6A
  FOLDFE: Foldfe
  FOLFD: Folfd
  FOLAC: Folac
  VITB12: Vitb12
  VITD: Vitd
  TOCPHA: Tocpha
  VITK1: Vitk1
  WATER: Water
}
// ...
export interface TotalDaily {
  ENERC_KCAL: EnercKcal2
  FAT: Fat2
  FASAT: Fasat2
  CHOCDF: Chocdf2
  FIBTG: Fibtg2
  PROCNT: Procnt2
  CHOLE: Chole2
  NA: Na2
  CA: Ca2
  MG: Mg2
  K: K2
  FE: Fe2
  ZN: Zn2
  P: P2
  VITA_RAE: VitaRae2
  VITC: Vitc2
  THIA: Thia2
  RIBF: Ribf2
  NIA: Nia2
  VITB6A: Vitb6A2
  FOLDFE: Foldfe2
  VITB12: Vitb122
  VITD: Vitd2
  TOCPHA: Tocpha2
  VITK1: Vitk12
}
// ...
export interface Digest {
  label: string
  tag: string
  schemaOrgTag?: string
  total: number
  hasRDI: boolean
  daily: number
  unit: string
  sub?: Sub[]
}

export interface Sub {
  label: string
  tag: string
  schemaOrgTag?: string
  total: number
  hasRDI: boolean
  daily: number
  unit: string
}

export interface Links2 {
  self: Self
}

export interface Self {
  href: string
  title: string
}